Implement delete transaction confirmation modal and update delete view logic

+++ transactions/views.py
@@ -100,7 +100,8 @@ def transaction_edit(request, transaction_id):
 def transaction_delete(request, transaction_id):
     qs = Transaction.objects.filter(owner=request.user)
     transaction = get_object_or_404(qs, id=transaction_id)
-    
-    transaction.delete()
-    messages.add_message(request, messages.SUCCESS, 'Transaction deleted!')
-    return redirect('transaction')  
+    if request.method == "POST":
+        transaction.delete()
+        messages.add_message(request, messages.SUCCESS, 'Transaction deleted!')
+        return redirect('transaction')  
+    return redirect('transaction')
